Re run config for a given window
