+++ scripts/constants.py
@@ -4,3 +4,8 @@ import enum
 class SplitDirection(enum.StrEnum):
     horizontal = "horizontal"
     vertical = "vertical"
+
+
+class WindowNamingOption(enum.StrEnum):
+    create = "create"
+    rename = "rename"
